# Conflicts:
#	files/assets/js/javascript.js
#	files/classes/user.py
#	files/templates/comments.html
#	files/templates/header.html
#	files/templates/submission.html
#	files/templates/submission_listing.html

+++ files/routes/oauth.py
@@ -9,8 +9,8 @@ from files.__main__ import app
 @app.get("/authorize")
 @auth_required
 def authorize_prompt(v):
-	client_id = request.args.get("client_id")
-	application = g.db.query(OauthApp).filter_by(client_id=client_id).first()
+	client_id = request.values.get("client_id")
+	application = g.db.query(OauthApp).options(lazyload('*')).filter_by(client_id=client_id).first()
 	if not application: return {"oauth_error": "Invalid `client_id`"}, 401
 	return render_template("oauth.html", v=v, application=application)
 
@@ -20,8 +20,8 @@ def authorize_prompt(v):
 @validate_formkey
 def authorize(v):
 
-	client_id = request.form.get("client_id")
-	application = g.db.query(OauthApp).filter_by(client_id=client_id).first()
+	client_id = request.values.get("client_id")
+	application = g.db.query(OauthApp).options(lazyload('*')).filter_by(client_id=client_id).first()
 	if not application: return {"oauth_error": "Invalid `client_id`"}, 401
 	access_token = secrets.token_urlsafe(128)[:128]
 	new_auth = ClientAuth(
@@ -32,6 +32,8 @@ def authorize(v):
 
 	g.db.add(new_auth)
 
+	g.db.commit()
+
 	return redirect(f"{application.redirect_uri}?token={access_token}")
 
 
@@ -40,15 +42,17 @@ def authorize(v):
 def request_api_keys(v):
 
 	new_app = OauthApp(
-		app_name=request.form.get('name'),
-		redirect_uri=request.form.get('redirect_uri'),
+		app_name=request.values.get('name'),
+		redirect_uri=request.values.get('redirect_uri'),
 		author_id=v.id,
-		description=request.form.get("description")[:256]
+		description=request.values.get("description")[:256]
 	)
 
 	g.db.add(new_app)
 
-	send_admin(NOTIFICATIONS_ACCOUNT, f"@{v.username} has requested API keys for `{request.form.get('name')}`. You can approve or deny the request [here](/admin/apps).")
+	send_admin(NOTIFICATIONS_ACCOUNT, f"{v.username} has requested API keys for `{request.values.get('name')}`. You can approve or deny the request [here](/admin/apps).")
+
+	g.db.commit()
 
 	return redirect('/settings/apps')
 
@@ -59,12 +63,14 @@ def request_api_keys(v):
 def delete_oauth_app(v, aid):
 
 	aid = int(aid)
-	app = g.db.query(OauthApp).filter_by(id=aid).first()
+	app = g.db.query(OauthApp).options(lazyload('*')).filter_by(id=aid).first()
 
-	for auth in g.db.query(ClientAuth).filter_by(oauth_client=app.id).all():
+	for auth in g.db.query(ClientAuth).options(lazyload('*')).filter_by(oauth_client=app.id).all():
 		g.db.delete(auth)
 
 	g.db.delete(app)
+
+	g.db.commit()
 
 	return redirect('/apps')
 
@@ -75,13 +81,15 @@ def delete_oauth_app(v, aid):
 def edit_oauth_app(v, aid):
 
 	aid = int(aid)
-	app = g.db.query(OauthApp).filter_by(id=aid).first()
+	app = g.db.query(OauthApp).options(lazyload('*')).filter_by(id=aid).first()
 
-	app.redirect_uri = request.form.get('redirect_uri')
-	app.app_name = request.form.get('name')
-	app.description = request.form.get("description")[:256]
+	app.redirect_uri = request.values.get('redirect_uri')
+	app.app_name = request.values.get('name')
+	app.description = request.values.get("description")[:256]
 
 	g.db.add(app)
+
+	g.db.commit()
 
 	return redirect('/settings/apps')
 
@@ -91,7 +99,7 @@ def edit_oauth_app(v, aid):
 @validate_formkey
 def admin_app_approve(v, aid):
 
-	app = g.db.query(OauthApp).filter_by(id=aid).first()
+	app = g.db.query(OauthApp).options(lazyload('*')).filter_by(id=aid).first()
 	user = app.author
 
 	app.client_id = secrets.token_urlsafe(64)[:64]
@@ -108,6 +116,8 @@ def admin_app_approve(v, aid):
 
 	send_notification(NOTIFICATIONS_ACCOUNT, user, f"Your application `{app.app_name}` has been approved. Here's your access token: `{access_token}`\nPlease check the guide [here](/api) if you don't know what to do next.")
 
+	g.db.commit()
+
 	return {"message": f"{app.app_name} approved"}
 
 
@@ -116,14 +126,15 @@ def admin_app_approve(v, aid):
 @validate_formkey
 def admin_app_revoke(v, aid):
 
-	app = g.db.query(OauthApp).filter_by(id=aid).first()
+	app = g.db.query(OauthApp).options(lazyload('*')).filter_by(id=aid).first()
+	if app.id:
+		for auth in g.db.query(ClientAuth).options(lazyload('*')).filter_by(oauth_client=app.id).all(): g.db.delete(auth)
 
-	for auth in g.db.query(ClientAuth).filter_by(oauth_client=app.id).all(): g.db.delete(auth)
+		send_notification(NOTIFICATIONS_ACCOUNT, app.author, f"Your application `{app.app_name}` has been revoked.")
 
-	g.db.flush()
-	send_notification(NOTIFICATIONS_ACCOUNT, app.author, f"Your application `{app.app_name}` has been revoked.")
+		g.db.delete(app)
 
-	g.db.delete(app)
+		g.db.commit()
 
 	return {"message": f"App revoked"}
 
@@ -133,14 +144,15 @@ def admin_app_revoke(v, aid):
 @validate_formkey
 def admin_app_reject(v, aid):
 
-	app = g.db.query(OauthApp).filter_by(id=aid).first()
+	app = g.db.query(OauthApp).options(lazyload('*')).filter_by(id=aid).first()
 
-	for auth in g.db.query(ClientAuth).filter_by(oauth_client=app.id).all(): g.db.delete(auth)
+	for auth in g.db.query(ClientAuth).options(lazyload('*')).filter_by(oauth_client=app.id).all(): g.db.delete(auth)
 
-	g.db.flush()
 	send_notification(NOTIFICATIONS_ACCOUNT, app.author, f"Your application `{app.app_name}` has been rejected.")
 
 	g.db.delete(app)
+
+	g.db.commit()
 
 	return {"message": f"App rejected"}
 
@@ -156,7 +168,7 @@ def admin_app_id(v, aid):
 			OauthApp.author)).filter_by(
 		id=aid).first()
 
-	pids=oauth.idlist(page=int(request.args.get("page",1)),
+	pids=oauth.idlist(page=int(request.values.get("page",1)),
 		)
 
 	next_exists=len(pids)==101
@@ -182,7 +194,7 @@ def admin_app_id_comments(v, aid):
 			OauthApp.author)).filter_by(
 		id=aid).first()
 
-	cids=oauth.comments_idlist(page=int(request.args.get("page",1)),
+	cids=oauth.comments_idlist(page=int(request.values.get("page",1)),
 		)
 
 	next_exists=len(cids)==101
@@ -215,11 +227,13 @@ def reroll_oauth_tokens(aid, v):
 
 	aid = aid
 
-	a = g.db.query(OauthApp).filter_by(id=aid).first()
+	a = g.db.query(OauthApp).options(lazyload('*')).filter_by(id=aid).first()
 
 	if a.author_id != v.id: abort(403)
 
 	a.client_id = secrets.token_urlsafe(64)[:64]
 	g.db.add(a)
+
+	g.db.commit()
 
 	return {"message": "Client ID Rerolled", "id": a.client_id}
